feat: working border countries redirect
